SumOfAllTheFactors

+++ src/Loops/Factorial.java
@@ -6,7 +6,7 @@ public class Factorial {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        double factorial = 1;
+        long factorial = 1;
         for (int i = 1; i <= n ; i++) {
             factorial *= i;
         }
